zone_controller_integration/requirements-test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.9.0
pytest-mock>=3.10.0

